agents split in separate files, common code
