+++ lib/agents/console.js
@@ -1,0 +1,8 @@
+// Generated by CoffeeScript 2.0.0-beta4
+var ConsoleAgent;
+module.exports = ConsoleAgent = {
+  enable: function (cb) {
+    console.log('Console#enable');
+    return cb(null, true);
+  }
+};
