+++ lib/inspector/socket_channel.js
@@ -5,22 +5,12 @@ debug = require('../debug-client');
 agents = require('../agents');
 _ = require('underscore');
 entryScript = require('../forked/entry_script');
-wrapperObject = function (type, description, hasChildren, frame, scope, ref, subtype) {
-  return {
-    type: type,
-    subtype: subtype,
-    description: description,
-    hasChildren: type === 'function' || type === 'object',
-    objectId: type === 'function' || type === 'object' ? '' + frame + ':' + scope + ':' + ref : null
-  };
-};
-logAndReturn = function (expr) {
-  console.log(expr);
-  return expr;
-};
+cache$ = require('../wrap_and_map');
+wrapperObject = cache$.wrapperObject;
+logAndReturn = cache$.logAndReturn;
 SocketChannel = function () {
   function SocketChannel(param$) {
-    var cache$, instance$, this$, this$1, this$2, this$3;
+    var cache$1, instance$, this$, this$1, this$2, this$3;
     instance$ = this;
     this.sendParsedScripts = function (c) {
       return SocketChannel.prototype.sendParsedScripts.apply(instance$, arguments);
@@ -29,10 +19,10 @@ SocketChannel = function () {
       return SocketChannel.prototype.onPauseOrBreakpoint.apply(instance$, arguments);
     };
     {
-      cache$ = param$;
-      this.socketConnection = cache$.socketConnection;
-      this.httpServer = cache$.httpServer;
-      this.socketServer = cache$.socketServer;
+      cache$1 = param$;
+      this.socketConnection = cache$1.socketConnection;
+      this.httpServer = cache$1.httpServer;
+      this.socketServer = cache$1.socketServer;
     }
     console.log('[bugger] SocketChannel created');
     this.hiddenFilePatterns = [];
@@ -74,11 +64,11 @@ SocketChannel = function () {
     return this.dispatchEvent('Timeline.eventRecorded', { record: record });
   };
   SocketChannel.prototype.handleRequest = function (msg) {
-    var agentName, args, cache$, functionName, this$;
+    var agentName, args, cache$1, functionName, this$;
     if (msg.method) {
-      cache$ = msg.method.split('.');
-      agentName = cache$[0];
-      functionName = cache$[1];
+      cache$1 = msg.method.split('.');
+      agentName = cache$1[0];
+      functionName = cache$1[1];
       console.log('[agents.handleRequest] ' + agentName + '#' + functionName);
       args = [];
       if (msg.params)
