+++ src/agents/console.coffee
@@ -1,0 +1,5 @@
+
+module.exports = ConsoleAgent =
+  enable: (cb) ->
+    console.log 'Console#enable'
+    cb(null, true)
